feat : 최단거리 합 계산

+++ src/main/java/com/team6/sole/domain/home/entity/Course.java
@@ -96,7 +96,7 @@ public class Course extends BaseTimeEntity {
     @Builder
     public Course(Long courseId, String thumbnailUrl, int scrapCount,
                   String title, String description, Date startDate,
-                  int duration, int distance,
+                  int duration, double distance,
                   Set<PlaceCategory> placeCategories, Set<WithCategory> withCategories, Set<TransCategory> transCategories,
                   Member writer, Member member, List<Place> places, List<CourseMember> courseMembers,
                   List<Declaration> declarations) {
